Moved org-clusters related REST API tests into org_clusters.go
